refactor/improve reliability
